requirements.txt
Django==2.2.13
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
